refactor: 패키지명 수정 inferfaces -> interfaces
- 관련 import 문 및 참조 경로 업데이트
