--- src/main/java/kr/hhplus/be/server/inferfaces/payment/dto/PaymentResponse.java
@@ -1,16 +1,0 @@
-package kr.hhplus.be.server.inferfaces.payment.dto;
-
-import kr.hhplus.be.server.domain.order.OrderStatus;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-
-public record PaymentResponse(
-        Long paymentId,
-        Long orderId,
-        Long userId,
-        long amount,
-        LocalDateTime createdAt,
-        OrderStatus orderStatus
-) {
-}
