+++ src/main/java/kr/hhplus/be/server/interfaces/coupon/dto/UserCouponRequest.java
@@ -1,0 +1,12 @@
+package kr.hhplus.be.server.interfaces.coupon.dto;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+
+public record UserCouponRequest(
+        @Schema(description = "쿠폰 ID", example = "123")
+        Long couponId,
+
+        @Schema(description = "사용자 ID", example = "1")
+        Long userId
+) {
+}
