+++ src/main/java/kr/hhplus/be/server/interfaces/order/OrderController.java
@@ -1,0 +1,31 @@
+package kr.hhplus.be.server.interfaces.order;
+
+import kr.hhplus.be.server.application.order.OrderFacade;
+import kr.hhplus.be.server.application.order.OrderResult;
+import kr.hhplus.be.server.interfaces.order.dto.OrderCreateRequest;
+import kr.hhplus.be.server.interfaces.order.dto.OrderResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1/orders")
+public class OrderController implements OrderControllerDocs {
+
+    private final OrderFacade orderFacade;
+
+    public OrderController(OrderFacade orderFacade) {
+        this.orderFacade = orderFacade;
+    }
+
+    @PostMapping
+    public ResponseEntity<OrderResponse> createOrder(
+            @RequestBody OrderCreateRequest request
+    ) {
+        OrderResult result = orderFacade.order(request.toCriteria());
+
+        return ResponseEntity.ok(OrderResponse.from(result));
+    }
+}
