+++ src/main/java/kr/hhplus/be/server/interfaces/payment/dto/PaymentResponse.java
@@ -1,0 +1,15 @@
+package kr.hhplus.be.server.interfaces.payment.dto;
+
+import kr.hhplus.be.server.domain.order.OrderStatus;
+
+import java.time.LocalDateTime;
+
+public record PaymentResponse(
+        Long paymentId,
+        Long orderId,
+        Long userId,
+        long amount,
+        LocalDateTime createdAt,
+        OrderStatus orderStatus
+) {
+}
